+++ src/System/Router/Router.php
@@ -7,6 +7,12 @@ class Router
   private static $routes = Array();
   private static $pathNotFound = null;
   private static $methodNotAllowed = null;
+  /** @var Contoller */
+  private static $contoller;
+
+  public static function setConttoller(Controller $contoller) {
+    self::$contoller = $contoller;
+  }
 
   /**
    * Short hand to readable regex url
@@ -86,8 +92,8 @@ class Router
 
   public static function view(string $uri, string $view_name, array $portal = [])
   {
-    self::match('get', $uri,
-      fn() => Controller::renderView($view_name, $portal)
+    return self::match('get', $uri,
+      fn() => self::$contoller::renderView($view_name, $portal)
     );
   }