+++ tests/Router/sample/sample.php
@@ -1,0 +1,1 @@
+<html><head></head><body><h1><?= $content->say ?></h1></body></html>